Modifications based on comments
