+++ caravel/assets/javascripts/explorev2/index.jsx
@@ -5,7 +5,6 @@ import ExploreViewContainer from './components/ExploreViewContainer';
 import { createStore, applyMiddleware, compose } from 'redux';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
-import { getDevEnhancer } from '../../utils/common';
 
 import { initialState } from './stores/store';
 
@@ -13,9 +12,6 @@ const exploreViewContainer = document.getElementById('js-explore-view-container'
 const bootstrapData = JSON.parse(exploreViewContainer.getAttribute('data-bootstrap'));
 
 import { exploreReducer } from './reducers/exploreReducer';
-
-const metrics = [];
-metrics.push(bootstrapData.viz.form_data.metric);
 
 const bootstrappedState = Object.assign(initialState, {
   datasources: bootstrapData.datasources,
@@ -26,14 +22,14 @@ const bootstrappedState = Object.assign(initialState, {
   vizType: bootstrapData.viz.form_data.viz_type,
   timeColumn: bootstrapData.viz.form_data.granularity_sqla,
   timeGrain: bootstrapData.viz.form_data.time_grain_sqla,
-  metrics: metrics.map((m) => ({ value: m, label: m })),
+  metrics: [bootstrapData.viz.form_data.metric].map((m) => ({ value: m, label: m })),
   since: bootstrapData.viz.form_data.since,
   until: bootstrapData.viz.form_data.until,
   havingClause: bootstrapData.viz.form_data.having,
   whereClause: bootstrapData.viz.form_data.where,
 });
 const store = createStore(exploreReducer, bootstrappedState,
-  compose(applyMiddleware(thunk), getDevEnhancer())
+  compose(applyMiddleware(thunk))
 );
 
 ReactDOM.render(
