+++ caravel/assets/utils/common.js
@@ -39,34 +39,3 @@ export function getDevEnhancer() {
   }
   return enhancer;
 }
-
-export const VIZ_TYPES = [
-  { value: 'dist_bar', label: 'Distribution - Bar Chart', requiresTime: false },
-  { value: 'pie', label: 'Pie Chart', requiresTime: false },
-  { value: 'line', label: 'Time Series - Line Chart', requiresTime: true },
-  { value: 'bar', label: 'Time Series - Bar Chart', requiresTime: true },
-  { value: 'compare', label: 'Time Series - Percent Change', requiresTime: true },
-  { value: 'area', label: 'Time Series - Stacked', requiresTime: true },
-  { value: 'table', label: 'Table View', requiresTime: false },
-  { value: 'markup', label: 'Markup', requiresTime: false },
-  { value: 'pivot_table', label: 'Pivot Table', requiresTime: false },
-  { value: 'separator', label: 'Separator', requiresTime: false },
-  { value: 'word_cloud', label: 'Word Cloud', requiresTime: false },
-  { value: 'treemap', label: 'Treemap', requiresTime: false },
-  { value: 'cal_heatmap', label: 'Calendar Heatmap', requiresTime: true },
-  { value: 'box_plot', label: 'Box Plot', requiresTime: false },
-  { value: 'bubble', label: 'Bubble Chart', requiresTime: false },
-  { value: 'big_number', label: 'Big Number with Trendline', requiresTime: false },
-  { value: 'bubble', label: 'Bubble Chart', requiresTime: false },
-  { value: 'histogram', label: 'Histogram', requiresTime: false },
-  { value: 'sunburst', label: 'Sunburst', requiresTime: false },
-  { value: 'sankey', label: 'Sankey', requiresTime: false },
-  { value: 'directed_force', label: 'Directed Force Layout', requiresTime: false },
-  { value: 'world_map', label: 'World Map', requiresTime: false },
-  { value: 'filter_box', label: 'Filter Box', requiresTime: false },
-  { value: 'iframe', label: 'iFrame', requiresTime: false },
-  { value: 'para', label: 'Parallel Coordinates', requiresTime: false },
-  { value: 'heatmap', label: 'Heatmap', requiresTime: false },
-  { value: 'horizon', label: 'Horizon', requiresTime: false },
-  { value: 'mapbox', label: 'Mapbox', requiresTime: false },
-];
