+++ caravel/views.py
@@ -1213,7 +1213,6 @@ class Caravel(BaseCaravelView):
                     "datasources": [(d.id, d.full_name) for d in datasources],
                     "datasource_id": datasource_id,
                     "datasource_type": datasource_type,
-                    "slice": slc.to_json(),
                     "user_id": g.user.get_id() if g.user else None,
                     "viz": json.loads(viz_obj.get_json())
                 }
